refactor(core): extract createPromptModePreset
The 3 small prompt-mode-staging helpers (preparePromptMode,
activatePromptMode, setCopyStartPosition) are now co-located in a
focused module. They were previously inline in init() and shared by
the activation-key handlers, the prompt-mode UI handlers, and the
action-context builder — co-locating them prevents subtle divergence
between callsites that all stage prompt-mode entry.

Folded the setCopyStartPosition wrapper into preparePromptMode since
they were a 1-call indirection.

index.tsx: 719 -> 710 (-9). 21 prompt-mode specs green.
